// madnessfeed-client/src/app/types.ts
export type ReadingMode = "all" | "unread"

export type ReadingOrder = "asc" | "desc"

export type ViewMode = "title" | "cozy" | "detailed" | "expanded"

export type ScrollMode = "always" | "never" | "if_needed"

export type IconDisplayMode = "always" | "never" | "on_desktop" | "on_mobile"
// ...
export interface Entry {
    id: string
    guid: string
    title: string
    content: string
    categories?: string
    rtl: boolean
    author?: string
    enclosureUrl?: string
    enclosureType?: string
    mediaDescription?: string
    mediaThumbnailUrl?: string
    mediaThumbnailWidth?: number
    mediaThumbnailHeight?: number
    date: number
    insertedDate: number
    feedId: string
    feedName: string
    feedUrl: string
    feedLink: string
    iconUrl: string
    url: string
    read: boolean
    starred: boolean
    markable: boolean
    tags: string[]
}

export interface Entries {
    name: string
    message?: string
    errorCount: number
    feedLink: string
    timestamp: number
    hasMore: boolean
    offset?: number
    limit?: number
    entries: Entry[]
    ignoredReadStatus: boolean
}

export interface FeedInfo {
    url: string
    title: string
}

export interface FeedInfoRequest {
    url: string
}

export interface FeedModificationRequest {
    id: number
    name?: string
    categoryId?: string
    position?: number
    filter?: string
    pushNotificationsEnabled: boolean
    autoMarkAsReadAfterDays?: number
}

export interface GetEntriesRequest {
    id: string
    readType?: ReadingMode
    newerThan?: number
    order?: ReadingOrder
    keywords?: string
    excludedSubscriptionIds?: string
    tag?: string
}

export interface GetEntriesPaginatedRequest extends GetEntriesRequest {
    offset: number
    limit: number
}

export interface PublicPageSettings {
    id?: number
    name?: string
    enabled: boolean
    showUncategorized: boolean
    token?: string
    categoryIds: number[]
}

export interface PublicSubscription {
    id: number
    name: string
    feedLink: string
    iconUrl: string
}

export interface PublicCategory {
    id: string
    name: string
    pageName?: string
    children: PublicCategory[]
    feeds: PublicSubscription[]
}
// ...
export interface SharingSettings {
    email: boolean
    gmail: boolean
    facebook: boolean
    twitter: boolean
    tumblr: boolean
    instapaper: boolean
    buffer: boolean
}

export type PushNotificationType = "ntfy" | "gotify" | "pushover"

export interface PushNotificationSettings {
    type?: PushNotificationType
    serverUrl?: string
    userId?: string
    userSecret?: string
    topic?: string
}

export interface Settings {
    language?: string
    readingMode: ReadingMode
    readingOrder: ReadingOrder
    showRead: boolean
    scrollMarks: boolean
    customCss?: string
    customJs?: string
    scrollSpeed: number
    scrollMode: ScrollMode
    entriesToKeepOnTopWhenScrolling: number
    starIconDisplayMode: IconDisplayMode
    externalLinkIconDisplayMode: IconDisplayMode
    markAllAsReadConfirmation: boolean
    markAllAsReadNavigateToNextUnread: boolean
    customContextMenu: boolean
    mobileFooter: boolean
    unreadCountTitle: boolean
    unreadCountFavicon: boolean
    disablePullToRefresh: boolean
    primaryColor?: string
    sharingSettings: SharingSettings
    pushNotificationSettings: PushNotificationSettings
}